This supersedes #62 with a better approach to decreasing invalidations. Instead of removing methods for converting `StaticInt` to `Int` this subtypes `Number`. This lets us keep a lot of the basic math methods at the loss of some generic index like behavior, (which we were losing anyway when removing natural conversion to `Int`).

~~This gets rid of __ALL__ invalidations on v1.7 and leaves the following on v1.8 and v1.9~~
<details>
<summary>SnoopCompile output</summary>

```julia
 inserting UnitRange{T}(start::Static.StaticNumber, stop) where T<:Real in Static at /Users/zchristensen/projects/Static-proposals/Static.jl/src/Static.jl:300 invalidated:
   mt_backedges: 1: signature Tuple{Type{UnitRange{_A}} where _A, Any, Any} triggered MethodInstance for UnitRange{T}(::Integer, ::Integer) where T<:Real (0 children)
                 2: signature Tuple{Type{UnitRange{_A}} where _A, Any, Any} triggered MethodInstance for UnitRange{T}(::Any, ::Any) where T<:Real (3 children)
   35 mt_cache

 inserting ifelse(::False, x, y) in Static at /Users/zchristensen/projects/Static-proposals/Static.jl/src/Static.jl:85 invalidated:
   mt_backedges: 1: signature Tuple{typeof(ifelse), Any, Real, Real} triggered MethodInstance for min(::T, ::T) where T<:Real (0 children)
                 2: signature Tuple{typeof(ifelse), Any, AbstractFloat, AbstractFloat} triggered MethodInstance for Base._fast(::typeof(max), ::AbstractFloat, ::AbstractFloat) (0 children)
                 3: signature Tuple{typeof(ifelse), Any, AbstractFloat, AbstractFloat} triggered MethodInstance for Base._fast(::typeof(min), ::AbstractFloat, ::AbstractFloat) (0 children)
                 4: signature Tuple{typeof(ifelse), Any, Int64, Int64} triggered MethodInstance for filter!(::Function, ::Vector{Union{Nothing, Base.UUID}}) (0 children)
                 5: signature Tuple{typeof(ifelse), Any, Int64, Int64} triggered MethodInstance for filter!(::Function, ::Vector{Base.UUID}) (0 children)
                 6: signature Tuple{typeof(ifelse), Any, Int64, Int64} triggered MethodInstance for filter!(::Function, ::Vector{Symbol}) (0 children)
                 7: signature Tuple{typeof(ifelse), Any, Int64, Int64} triggered MethodInstance for filter!(::Base.var"#37#39", ::Vector{Any}) (8 children)
   3 mt_cache
```

</details>

~~I've already made a PR where I tracked down the source of these in to poor inference when filtering when running Pkg. (https://github.com/JuliaLang/julia/pull/45452)~~

EDIT: PR accepted

Here are the load times:
* v1.7
  - pr: 0.028802 seconds (42.68 k allocations: 2.577 MiB, 19.91% compilation time)
  - master: 0.049850 seconds (89.03 k allocations: 5.158 MiB, 9.87% compilation time)
* v1.8
  - pr: 0.029359 seconds (46.13 k allocations: 2.806 MiB, 17.48% compilation time)
    - still has invalidations now fixed in master 
  - master: 0.058451 seconds (100.29 k allocations: 5.878 MiB, 8.20% compilation time)
* v1.9
  - pr: 0.033590 seconds (46.10 k allocations: 2.806 MiB)
  - master: 0.032208 seconds (89.56 k allocations: 5.784 MiB)

These are updated after [a70c8e3](https://github.com/SciML/Static.jl/pull/64/commits/a70c8e32242ce373c837fec8aebb6ca8f3e1395e).
I'm not working on a computer that's great for benchmarking and the time in seconds isn't very accurate, so it's probably more useful to compare allocations between this PR and master