@ChrisRackauckas 2022-05-30
This does seem like a pretty good solution.