@Tokazama 2022-05-31
I'll see if I can preemptively reduce some of these errors in ArrayInterface