@chriselrod 2022-06-02
Mind bumping the minor version, to indicate this is a breaking release?