@Tokazama 2022-06-02
There are some things I still don't have a solution to here. For example, passing StaticInt to CartesianIndex requires a new method but that creates over 100 invalidations.