@chriselrod 2022-06-02
Which function is invalidated?
Is it deep enough (i.e. not something like `eachindex`) that we can just define `StaticCartesianIndex` as an alternative?