@Tokazama 2022-06-02
I came across the issue with [this line](https://github.com/JuliaSIMD/CloseOpenIntervals.jl/blob/1e31677e43f02dfebb75c243318f9ef9b84e84d7/test/runtests.jl#L4). It's because `AbstractCloseOpen` returns a `StaticInt` for `first`. It's probably easy to fix in that one package. I'm not sure if it will keep coming up though.