@cscherrer 2022-06-02
Thanks @Tokazama. Getting rid of the invalidations is great! I guess this will now have some of the same challenges as Symbolics, for example
```julia
julia> @variables i::Int
1-element Vector{Num}:
 i

julia> i isa Integer
false
```
The biggest concern here is that people get very used to writing methods like `f(i::Integer)`. With Symbolics and now Static going in this direction, there needs to be a more flexible way to write things like this.