@ChrisRackauckas 2022-06-02
Can you add a bunch of downstream tests? If those pass, then I think we're ready to rip off the bandaid.