@Tokazama 2022-06-02
> Can you add a bunch of downstream tests? If those pass, then I think we're ready to rip off the bandaid.

Right now I know that `LoopVectorization` will fail because it has structs with fields that specify an `Integer` subtype. I'm trying to do a quick sweep of that stuff and then I'll add some downstream tests