@chriselrod 2022-06-03
> There are some things I still don't have a solution to here. For example, passing StaticInt to CartesianIndex requires a new method but that creates over 100 invalidations.

Hmm. It is important that this
https://github.com/JuliaSIMD/LoopVectorization.jl/blob/cbb9f2324ec3cb2c4aa5523d8c4fb98147aa7855/test/offsetarrays.jl#L99
still passes its static size information to LoopVectorization.
It makes a big difference to both compile and runtimes.