@Tokazama 2022-06-03
> Perhaps I should add a test that CartesianIndices returns an object with static size information.

Do you mean the value returned from iterating `CartesianIndices` needs to be able to contain static info?