@Tokazama 2022-06-03
`CartesianIndices` are only supposed to return `CartesianIndex`. It's just not strictly enforced. I've been playing around with an `Indices` type that acts more like Dex's index set types and carries bounds checking information. This particular issue might be a good reason to seriously consider implementing it.