@Tokazama 2022-06-05
> It seems to me `Irrational` is pretty close to the way we'd want `StaticFloat64` to be set up:
> 
> ```julia
> julia> Irrational{:π} <: Real
> true
> ```
> 
> I don't know of any problems this leads to (are there any?). Could this design could be adapted for `StaticFloat64`, so we could keep the current subtyping?

I'm not sure I understand. Are you suggesting that we assign a symbol as the static value and then use dispatch to derive the float value, like...

```julia
struct StaticFloat{sym} <: AbstractFloat end

Base.float(::StaticFloat{:float_one}) = 1.0

```