@cscherrer 2022-06-05
> Are you suggesting that we assign a symbol as the static value and then use dispatch to derive the float value

No, I think the type parameter should be set up to map easily to the "dynamic" value, so in this case just a `Float64`. 

I could be missing something, but to me this seems like evidence the problem can be solved without sacrificing the type hierarchy.