@Tokazama 2022-06-05
> > The subtyping is what causes invalidations, not where we store the values.
> 
> I see, so it's not type instability but invalidation, and irrational avoid this problem because they're in Base. Then the design is irrelevant to the problem, and the "just put it in Base" solution is (sadly) not getting much traction. I think I see now, thank you.

Exactly